PYTHON  FAST API/requirements.txt
# FastAPI and core dependencies
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7

# Database - DynamoDB
boto3==1.35.80
botocore==1.35.80

# Environment configuration
python-dotenv==1.1.1
